dashboard/backend/requirements.txt
# Flask web framework
Flask>=3.0.0,<4.0.0
Flask-CORS>=4.0.0,<5.0.0
Flask-Session>=0.5.0,<1.0.0

# Database
SQLAlchemy>=2.0.0,<3.0.0

# Image processing
Pillow>=10.0.0,<11.0.0

# HTTP requests (for thumbnail generation)
requests>=2.31.0,<3.0.0

# Environment variables
python-dotenv>=1.0.0,<2.0.0

# Testing
pytest>=7.4.0,<8.0.0
pytest-flask>=1.2.0,<2.0.0
pytest-cov>=4.1.0,<5.0.0

# Production server
gunicorn>=21.2.0,<22.0.0
